* commit '1b42439946fc6347ae192a503423f10acd8ba9cd':
  changing appearance of checkboxes of contracts list

+++ src/main/java/com/endava/service_system/dao/InvoiceDao.java
@@ -34,5 +34,5 @@ public interface InvoiceDao extends JpaRepository<Invoice,Long> {
             "join i.contract c " +
             "join c.company co " +
             "join co.credential cr where cr.username=:username and i.invoiceStatus='CREATED'")
-    int[] getAllInvoicesIdsByCompanyUsername(@Param("username") String authenticatedUsername);
+    long[] getAllInvoicesIdsByCompanyUsername(@Param("username") String authenticatedUsername);
 }
